Loop Toggle (kinda)
So, the toggle is *there*, and at one point I damn near had perfect
functionality. For the purpose of the app still working correctly I
have commented out code breaking stuff. I chose to leave the toggle, as
well as some functionality tied to whether the track is being played or
not, in there for now, as it just looks kinda cool. Kinda helps to fill
up the screen. Eventually I’m confident we can get this working, we
just don’t have the time now.